WIP basic model training in preprocessing - polish advanced cases
